requirements.txt
python-keystoneclient
python-glanceclient
python-novaclient

python-mysqldb